Extract `MaybeFunctions` and add in `Prelude`.
